requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-WTF>=1.1
Flask-Login>=0.6
python-dotenv>=1.0
psycopg2-binary>=2.9
Werkzeug>=2.2
Flask-HTTPAuth>=4.8
email-validator>=2.0
Flask-Cors>=3.0